% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{move}
\alias{move}
\title{Generic Function to Move Objects}
\usage{
move(rects, ...)
}
\arguments{
\item{rects}{objects to be moved}

\item{\dots}{arguments passed to class methods}
}
\description{
Generic Function to Move Objects
}
