% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{move.rects}
\alias{move.rects}
\title{Move Rectangles}
\usage{
\method{move}{rects}(
  rects,
  dx = 0,
  dy = 0,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  each = FALSE,
  ...
)
}
\arguments{
\item{rects}{A "rects" object}

\item{dx}{delta x}

\item{dy}{delta y}

\item{top}{top position}

\item{bottom}{bottom position}

\item{left}{left position}

\item{right}{right position}

\item{each}{logical indicating whether to move each rectangle or the group of
rectangles}

\item{\dots}{additional arguments (currently not used)}
}
\description{
Move Rectangles
}
