% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{new_rects}
\alias{new_rects}
\title{Create a "rects" object}
\usage{
new_rects(
  w = 1,
  h = 1,
  llx = 0,
  lly = 0,
  lbl_text = NULL,
  lbl_align = "centre",
  density = -1,
  angle = 45,
  col = NA,
  border = graphics::par("fg"),
  lty = graphics::par("lty"),
  lwd = graphics::par("lwd"),
  n = NULL
)
}
\arguments{
\item{w}{width(s) of rectangle(s)}

\item{h}{height(s) of rectangle(s)}

\item{llx}{x position(s) of lower left corner(s) of rectangle(s)}

\item{lly}{y position(s) of lower left corner(s) of rectangle(s)}

\item{lbl_text}{text label(s)}

\item{lbl_align}{label alignment(s), default: "centre"}

\item{density}{density of shade lines. Default: -1}

\item{angle}{angle of shade lines. Default: 45}

\item{col}{fill colour(s) of rectangles}

\item{border}{border colour(s) of rectangles}

\item{lty}{line type(s) of rectangles}

\item{lwd}{line width(s) of rectangles}

\item{n}{number of rectangles. All other arguments are recycled to vectors of
this length.}
}
\description{
Create a "rects" object
}
