% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{plot.rects}
\alias{plot.rects}
\title{Plot Rectangles}
\usage{
\method{plot}{rects}(x, add = !is.null(grDevices::dev.list()), cex = 1, y = NULL, ...)
}
\arguments{
\item{x}{object of class "rects"}

\item{add}{logical indicating whether to add rectangles to an existing plot
or to start a new plot}

\item{cex}{character expansion factor}

\item{y}{not used. Just there to comply with the generic plot() interface.}

\item{\dots}{additional arguments (currently not used)}
}
\description{
Plot Rectangles
}
