% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{separate.rects}
\alias{separate.rects}
\title{Separate Rectangles}
\usage{
\method{separate}{rects}(x, dx = 0, dy = 0, ...)
}
\arguments{
\item{x}{"rects" object}

\item{dx}{space in horizontal direction to be put in between the rectangles}

\item{dy}{space in vertical direction to be put in between the rectangles}

\item{\dots}{further arguments (currently not used)}
}
\description{
Separate Rectangles
}
