% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{stack}
\alias{stack}
\title{Stack Rectangles}
\usage{
stack(rects, ...)
}
\arguments{
\item{rects}{a "rects" object}

\item{\dots}{further arguments passed to \code{stack.rects}}
}
\description{
Stack Rectangles
}
