% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{stack.rects}
\alias{stack.rects}
\title{Stack Rectangles Vertically or Horizontally}
\usage{
\method{stack}{rects}(rects, horizontal = FALSE, delta = 0, reverse = FALSE, ...)
}
\arguments{
\item{rects}{a "rects" object as returned by \code{\link{new_rects}}}

\item{horizontal}{whether to stack the rectangles horizontally or not. The
default is \code{FALSE}, i.e. rectangles are stacked vertically.}

\item{delta}{space between rectangles, default: 0}

\item{reverse}{whether or not to reverse the stack order. The default is
\code{FALSE}.}

\item{\dots}{not used}
}
\description{
Stack Rectangles Vertically or Horizontally
}
\examples{
rects <- new_rects(w = 1:3)
stacked_vertically <- stack(rects)
stacked_horizontally <- stack(rects, horizontal = TRUE)
plot(stacked_vertically, add = FALSE)
plot(stacked_horizontally, add = FALSE)
stacked_vertically_spaced <- stack(rects, delta = 0.1)
stacked_horizontally_spaced <- stack(rects, horizontal = TRUE, delta = 0.1)
plot(stacked_vertically_spaced, add = FALSE)
plot(stacked_horizontally_spaced, add = FALSE)
}
